Refactor migrations
Move migrations to the separate package

Rename database/migrations to database/migrate

Improve migrations CLI

Move CLI from database/cli to database/migrate/cli

Move CLI actions to the separate file
